Synchronize confusion matrix

+++ src/metrics.py
@@ -10,7 +10,7 @@ from src.general import xywh2xyxy, box_iou
 class ConfusionMatrix:
     # Updated version of https://github.com/kaanakan/object_detection_confusion_matrix
     def __init__(self, nc, conf=0.25, iou_thres=0.45):
-        self.matrix = np.zeros((nc + 1, nc + 1))
+        self.matrix = np.zeros((nc + 1, nc + 1), dtype=np.int32)
         self.nc = nc  # number of classes
         self.conf = conf
         self.iou_thres = iou_thres
